requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.10
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-extensions==2.2.1
django-import-export==2.0.2
django-js-asset==1.2.2
django-perms==2.1.0
django-taggit==1.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
GDAL==2.4.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.6.3
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==0.6.1
numpy==1.18.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==6.2.0
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-social-auth==0.2.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
rest-social-auth==2.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==1.1.0
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
